trash还原问题修复 Signed-off-by: yangjingbo10 <yangjingbo10@huawei.com>

+++ interfaces/kits/native/trash/include/file_info.h
@@ -22,7 +22,7 @@ struct FileInfo {
     std::string uri;
     std::string srcPath;
     std::string fileName;
-    int64_t mode { 0 };
+    int32_t mode { 0 };
     int64_t size { 0 };
     int64_t mtime { 0 };
     int64_t ctime { 0 };
